fix(generic-oauth): address CI failures and review comments
- Fix cspell: rename EMAIL_DOESNT_MATCH to EMAIL_DOES_NOT_MATCH
- Fix discovery validation: relax RFC 8414 issuer check to only validate
  the issuer is a valid URL (the strict path comparison broke legitimate
  setups where discoveryUrl origin differs from the issuer, e.g. proxied
  or path-prefixed deployments)
- Fix RFC 9207: only validate iss when present in the callback (older
  OAuth servers that don't support RFC 9207 omit it entirely)
- Fix private_key_jwt migration: use storeClientSecret instead of raw
  generateRandomString to respect the configured storage method
- Move OAUTH_CALLBACK_ERROR_CODES to oauth2/error-codes.ts (core owns
  callback routing, not the plugin)
- Remove iss injection from test helper (no longer needed since missing
  iss is accepted)
