+++ packages/better-auth/src/plugins/generic-oauth/generic-oauth.test.ts
@@ -92,14 +92,6 @@ describe("oauth2", async () => {
 
 		if (!location) throw new Error("No redirect location found");
 
-		// RFC 9207: inject iss parameter into the callback URL
-		// (the mock server doesn't add it, but real OIDC servers do)
-		const callbackWithIss = new URL(location);
-		if (!callbackWithIss.searchParams.has("iss")) {
-			callbackWithIss.searchParams.set("iss", `http://localhost:${port}`);
-		}
-		location = callbackWithIss.toString();
-
 		let callbackURL = "";
 		let setCookieHeader = "";
 		const newHeaders = new Headers();
